rename asyncJS.js to caolan-async.js
