+++ test/caolan-async.js
@@ -1,23 +1,23 @@
 const Test = require('ava').test;
 
-const asyncJs = require('./../asyncJs');
+const caolanAsync = require('./../caolan-async');
 
 Test.cb('B is called after A is complete', t => {
-  asyncJs.run(undefined, (err, result) => {
+  caolanAsync.run(undefined, (err, result) => {
     t.true(result.callOrder.indexOf('B Start') > result.callOrder.indexOf('A Finish'));
     t.end();
   });
 });
 
 Test.cb('C is called after A is complete', t => {
-  asyncJs.run(undefined, (err, result) => {
+  caolanAsync.run(undefined, (err, result) => {
     t.true(result.callOrder.indexOf('C Start') > result.callOrder.indexOf('A Finish'));
     t.end();
   });
 });
 
 Test.cb('D is called after A and B are complete', t => {
-  asyncJs.run(undefined, (err, result) => {
+  caolanAsync.run(undefined, (err, result) => {
     t.true((result.callOrder.indexOf('D Start') > result.callOrder.indexOf('A Finish'))
        && ((result.callOrder.indexOf('D Start') > result.callOrder.indexOf('B Finish'))));
     t.end();
@@ -25,7 +25,7 @@ Test.cb('D is called after A and B are complete', t => {
 });
 
 Test.cb('E is called after A and C are complete', t => {
-  asyncJs.run(undefined, (err, result) => {
+  caolanAsync.run(undefined, (err, result) => {
     t.true((result.callOrder.indexOf('E Start') > result.callOrder.indexOf('A Finish'))
        && ((result.callOrder.indexOf('E Start') > result.callOrder.indexOf('C Finish'))));
     t.end();
@@ -33,7 +33,7 @@ Test.cb('E is called after A and C are complete', t => {
 });
 
 Test.cb('It calls each function as soon as it\'s dependencies are satisfied', t => {
-  asyncJs.run(undefined, (err, result) => {
+  caolanAsync.run(undefined, (err, result) => {
     t.deepEqual(result.callOrder, [
       'A Start',
       'A Finish',
